add client but password not coded
