+++ src/Entity/Client.php
@@ -4,7 +4,8 @@
 namespace App\Entity;
 
 
-use App\DTO\Client\CreateClientFromRequestInput;
+
+use App\DTO\Client\CreateClient\CreateClientFromRequestInput;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
