+++ src/Validator/Constraints/Client/Password/containsOneUpCase.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace App\Validator\Constraints\Client\Password;
+//todo: delete
+/**
+ * Class containsOneUpCase
+ * @package App\Validator\Constraints\Client\Password
+ * @Annotation
+ */
+class containsOneUpCase
+{
+    public $message='Password must have at least one Uppercase';
+}
